refactor: codebase
- new structure
- clear naming
- update CLI help message
